requirements.txt
flask
gunicorn
requests
psycopg2-binary
prometheus_client
